+++ test/smoke_test.rb
@@ -12,10 +12,11 @@ test "simple case" do
   expect("s").to_be_a String
   expect(123).to_be_kind_of Numeric
   expect(123).to_be_instance_of Integer
+  expect([]).to_be_empty
   expect({ a: 1 }).to_include :a
   expect(/^fo/).to_match "foo"
   expect(1..4).to_contain_exactly 1, 3, 4, 2
-  expect([]).to_be_empty
+  expect("あ").to_have_attributes size: 1, bytesize: 3
   expect { 1 / 0 }.to_raise ZeroDivisionError
   expect { 1 / 0 }.to_raise ZeroDivisionError, "divided"
   expect { 1 / 2 }.to_raise_nothing